django_project/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-modeltranslation==0.19.10
pillow==10.4.0
psycopg-binary==3.2.3
psycopg2-binary==2.9.9
sqlparse==0.5.1
